Enable dynamic linking
Doing this in Cargo.toml rather than as a CLI option to `cargo run`
because to force `rust-analyzer` to use dynamic linking too. Specifying
this on the CLI resulted in `rust-analyzer` recompiling the whole thing
without dynamic linking, blocking the 'real' compilation and then
thereby forcing the real compilation to recompile everything on a second
time.